moved student-specific UI tests to StudentUITests
