+++ Student/StudentUITests/People/PeopleTests.swift
@@ -18,6 +18,7 @@
 
 import XCTest
 import TestsFoundation
+@testable import CoreUITests
 
 enum CoursePeople {
     static func person(name: String) -> Element {
@@ -32,8 +33,6 @@ enum PersonContextCard {
 }
 
 class PeopleTests: CoreUITests {
-    override var abstractTestClass: CoreUITests.Type { return PeopleTests.self }
-
     func testCourseUsersAndUserContextCardDisplay() {
         // Dashboard
         Dashboard.courseCard(id: "262").waitToExist()
