+++ Student/StudentUITests/Quizzes/QuizTests.swift
@@ -18,10 +18,9 @@
 
 import XCTest
 import TestsFoundation
+@testable import CoreUITests
 
 class QuizTests: CoreUITests {
-    override var abstractTestClass: CoreUITests.Type { return QuizTests.self }
-
     func testQuizQuestionsOpenInNativeView() {
         Dashboard.courseCard(id: "263").waitToExist()
         Dashboard.courseCard(id: "263").tap()
