+++ Student/StudentUITests/Submissions/SubmissionDetails/SubmissionDetailsTests.swift
@@ -20,6 +20,7 @@ import Foundation
 @testable import Core
 import TestsFoundation
 import XCTest
+@testable import CoreUITests
 
 class SubmissionDetailsTests: StudentUITestCase {
     lazy var course: APICourse = {
